Rename files after deciding that we will keep an annual check as well as measurements
