+++ 2024-frontend/src/views/WasteMeasurementTunnel/WasteMeasurementSteps/BreakdownBySource.vue
