+++ 2024-frontend/src/views/WasteMeasurementTunnel/WasteMeasurementSteps/TotalWaste.vue
